particle refactoring
