+++ src/main/java/me/pugabyte/nexus/features/minigames/perks/particles/FlameParticle.java
@@ -1,13 +1,11 @@
 package me.pugabyte.nexus.features.minigames.perks.particles;
 
-import me.pugabyte.nexus.features.minigames.models.perks.PerkCategory;
-import me.pugabyte.nexus.features.minigames.models.perks.common.TickablePerk;
+import me.pugabyte.nexus.features.minigames.models.perks.common.ParticlePerk;
 import org.bukkit.Material;
 import org.bukkit.Particle;
-import org.bukkit.entity.Player;
 import org.bukkit.inventory.ItemStack;
 
-public class FlameParticle extends TickablePerk {
+public class FlameParticle extends ParticlePerk {
 	@Override
 	public String getName() {
 		return "Flame";
@@ -24,17 +22,17 @@ public class FlameParticle extends TickablePerk {
 	}
 
 	@Override
-	public PerkCategory getPerkCategory() {
-		return PerkCategory.PARTICLE;
-	}
-
-	@Override
 	public int getPrice() {
 		return 10;
 	}
 
 	@Override
-	public void tick(Player player) {
-		particle(player, Particle.FLAME, 0.002d, player.getWorld().getPlayers());
+	public Particle getParticle() {
+		return Particle.FLAME;
+	}
+
+	@Override
+	public double getSpeed() {
+		return 0.002d;
 	}
 }
